requirements.txt
numpy~=1.21.4
psycopg2-binary~=2.9.5
pandas~=1.3.5
beautifulsoup4~=4.11.2
requests~=2.28.2